Header Menu Spacing Bug FIXED
The fix is to change the default position value for the .stuckMenu
class. stickUp was setting the element to be positioned relative,
changing it to be fixed upon scrolling. I just made it fixed and kept
the 63px spacing. This renders the stickUp pretty useless, but w/e. For
now it’s fine. If I get bored I’ll remove it and stick the menu with
css.